requirements.txt
websockets
pyfiglet
python-dotenv